Remove debian/*prerm files, dangerous/no overrides
If an cluster operator needs to upgrade to a new Ceph release, it
may not be desireable to start/stop/restart daemons at the time of
package installation. There is not way to coax apt into ignoring
prerm scripts by way of Dpkg::Options::, the only option is to
manually alter /var/lib/dpkg/info/{ceph,radosgw,ceph-mds}.prerm to
have it return true.

See Debian Policy chapter 6.6, section 1, items 1 & 2 for details
about when prerm scripts are called:

http://www.debian.org/doc/debian-policy/ch-maintainerscripts.html

An example of why an operator might not want service handling tied
to package installation would be upgrading the packages on a large
cluster of machines first, then later performing rolling restarts.